Added credit for cover photo
Note: This forks `the-north-tower.svelte` away from the current notion2svelte output. So either future changes in Notion will need to be manually curated back into Svelte, or (and this is the eventual plan) _notion2svelte_ will need an update so that it can provide custom fields from Notion page props to `setContext` in the Svelte output.