Nuclear fix for the `%thispage%` gaffe
I hadn't meant to push that to the site. Was intended to first have Svelte auto-fill it with the page title. That's not my top dev priority at the moment, since what I really want is to develop what I've started with `setContext`. Decided, for now, to skip the big, complex email template & strip it down to essentials