White space hack
The trouble with using fixed-width font code blocks to lay out poetry is that the leading spaces end up compressed so that the indentations look like mistakes. This change forces the issue, but I haven't tested it in very many cases, so it may prove too brittle in the end. 🤞