Forced rename to fix capitalization discrepency
This is a weird one. Basically, it seems I originally checked in & pushed a lowercase 'intro.svelte', and since then, Git hasn't seen any changes, so when I changed the capitalization locally, it failed to catch it as a rename.

To make the current rename happen, I changed `Intro.svelte` to `Intros.svelte`, staged the change, then changed it back and staged _that_ change. I have a vague recollection of doing this previously, so…🤞